(PIE-302) Add initial acceptance tests
These will be iterated on in future commits.

The changes here also include some fixes to get the acceptance tests
working. This includes using the ini_subsetting resource to declare the
'servicenow' report processor, which cleans up much of the manual cruft
that was initially added and also fixes a previous error where the 'store'
report processor was accidentally getting included in the 'reports' setting.

Signed-off-by: Enis Inan <enis.inan@puppet.com>