(PIE-304) Remove PE console reports link and link to PE console instead
Fine-grained PE console links like the reports link are unstable even
for Y PE releases. For example, the link is different for PE 2019.2
and PE 2019.8. Thus, the best and most stable solution is to make a note
that the report can be viewed in the PE console, and add a link to the
PE console's main URL. This commit implements that solution.

Signed-off-by: Enis Inan <enis.inan@puppet.com>