(PIE-307) Fill-in the remaining incident fields
These are the incident fields that appear in ServiceNow's default
incident UI. Some things to note:

* caller_id is a required parameter because it is a required field in
incident creation.

* incident_state (used in the report2snow module) is not needed.
Based on
https://community.servicenow.com/community?id=community_question&sys_id=34caf657db2e534023f4a345ca96198c,
state is enough b/c there's rules syncing incident_state and state.

* severity (used in report2snow) doesn't show up in the default incident
UIs, so it is omitted until a customer asks for it.

* priority (set in report2snow) is typically calculated from impact and
urgency. Thus, it is also omitted until a customer asks for it.

* escalation (set in report2snow) is a legacy field according to https://community.servicenow.com/community?id=community_question&sys_id=3b604b87db3c674067a72926ca9619b1

* location (shown in one of the default UIs) seems a bit strange for a
Puppet report, so it is omitted until a customer asks for it.

Note that the UX here will likely change in the future to allow for more
fine-grained configuration (e.g. configure these fields on report status
vs. toplevel).

Signed-off-by: Enis Inan <enis.inan@puppet.com>