(PIE-299) Add support for oauth tokens.
Similar to how the cmdb integration works we will require either a
user/password or oauth_token but not both. Validation was added to
init.pp to ensure this.

The acceptance test mock server was already configured to handle Bearer
tokens.

The README still needs to be updated with instructions for how to use a
token, but that can be copied from the cmdb integration.

Tokens can be encrypted in the same was passwords are.