(PIE-306) Timestampt Too Verbose
This change formats the timestamp slightly better, and moves it to the
back of the short description.

The improved formatting should be the actual time displayed to the user
when the go to the `Reports` screen in the PE Console. This should
allow users searching for a particular report that caused an incident to
`ctrl-f` search the reports screen to find the correct one based on this
improved timestamp.

Moving the stamp tto back of the string allows the user to parse the
important information about severity and target first.

In the console, when the timezone is `UTC`, that value is replaced by
the string `Z`. Since that seems to be the most common time zone name
mapping I did that one manually, but any other time zones will get the
time zone abreviation value that Ruby has directly.