(PIE-311) Fine grain description of the incident creation reason
This adds a incident_description method which is used to populate the
incident description with report labels that describe why an incident
was created as well as the pe console reports url.

The create_incident? method has been replaced by  two helper methods,
calcualte_event_conditions, and calculate_satisfied_conditions which
don't change any functionality or ordering, but allow us to only
calculate the satisfied conditions once so that they can be reused by
the incident_description function.

Unit tests have been modified to ensure the description contains this
information.