(PIE-327) Add $pe_console_url Default Value
This change adds a default value for the `pe_console_url` parameter and
makes setting that value optional during class assignment.

The configuration value used is $settings::report_server
(https://puppet.com/docs/puppet/6.17/configuration.html#report_server)
Unfortunately, the rest of the URL for finding a node report is not
stable, so the name of the server and adding a default for this
parameter is the best we can do. We cannot put a full link to the final
report in the description of the incidents this module creates.

$settings::report_server will most likely be the correct choice for the
vast majority of users. If for some reason they are hosting reports on a
server that is not the same as their Master, Mast of Masters, or console
host, then the value is still configurable by the user.