(PIE-348) Add ignore OK events rule task
This change adds a task to the module that will add an event rule to
ignore events with a severity level of 'OK'. These events should
simply represent Puppet runs where nothing happened.

If anything like changes or failures happened, those should events
should have a different severity level that can be targeted with
different rules for how to create alerts for them.