(PIE-358) Events with no changes now send a severity
If an event status is not in one of the statuses returned by the
calculate_event_conditions method, a null status will be sent to
servicenow which creates an error on the servicenow end. This fix adds a
default status for no change statuses that is applied if a changed
status is not found. A test has been added with an absurd severity to
make sure there is no overlap with the other "1" severity statuses.