(PIE-352) Add Report Labels
This change adds report labels to the descriptions of both incidents and
events. The labels help users understand if what changes are represented
in a particular incident or report.

While the incident severity will only reflect the highest severity
change in the report, the labels will reflect all of the different
change types that may have been present in a given report.