(MAINT) Update PE Install Script
This change switches the install script to look at the more stable
`archives` area of Artifactory for downloading PE. This lays the
groundwork for turning the script into a task that can easily be
parameterizd for PE version and platform version.

In order to do matrix builds with this script it will either have to be
turned into a task with parameters, or edited on the build server before
being executed on the target server.