(PIE-372) Catalog Compilation Failures Not Handled
Previously when a puppet run resulted in a catalog failure, there were
no report labels populated in the report description.

This change ensures that if a catalog failure occurs there is a report
label is still generated, and additionally the log messages are
included. We include the log messages in the case of catalog failure
because otherwise there is no way to provide the user with feedback
on what might have caused the failure.

Previously the module would also not send the correct event severity
when catalog compilation failed. This is because event severity
selection was based entirely on the resource events. However, when the
catalog fails to compile there are no resource events, so the wrong
event severity would be selected. This change addresses that issue.