(PIE-379) Add disabled param
If an administrator removes the event management class from a node, that
would not be enough to get the report processor to stop sending events.
The admin would need to go into the puppet.conf and remove the report
processor from the configuration. Especially if a customer only wants to
stop sending events for a short period, this isn't great.

Instead we provide this parameter which will serve as an off switch to
be used as needed.