(PIE-383) Add HTTP Timeout Parameters
This change adds user parameters to set the HTTP timeout arguments
to the `Net::HTTP.start` method used by the module to make
Servicenow API requests.

The `http_read_timeout` parameter sets the following arguments to the
same value: `ssl_timeout`, `read_timeout`, `connect_timeout`. Most users
will consider these to be functionally the same timeout so there is
little reason at the momen to increase the module's complexity by
exposing each of those arguments as a separate parameter for users.

The `http_write_timeout` parameter sets the `write_timeout` argument.

These parameters are not explicitly acceptance tested because that is
functionally testing the `Net::HTTP` class, which is beyond the scope of
this module. They are unit tested to ensure we don't make any changes
that might accidentaly cause them not to be passed to the underlying
Ruby class.