(PIE-381) Reorder Description Field
When events and incidents are viewed in a table view in Servicenow, the
the first portion of the description is a static blurb about going to
the console to review the report that generated the table entry.

This change reorders that field so that the first information is the
change type tags. This gives the user at least a slightly helpful
bit of information before clicking into the full entry.