(PIE-395) Use Sensitive Datatypes
Switch to using Sensitive as the datatype for the password and oauth
token parameters.

This change requires working around a bug in the the rspec-puppet 2.7.10
gem wherin the sensitive.rb file is not downloaded with the rest of the
gem. As a result, a direct copy of sensitive.rb from the rspec-puppet
source code has been added to this repo at
spec/support/unit/sensitive.rb. This file should be removed in favor of
the rspec-puppet copy when the bug in that gem is fixed.