(PIE-409) Add change type hash keys
If a user wants to group events by the resource change events,
Servicenow has a bug where it cannot handle the raw resource data that
we add to the additional info field. Some sort of problem is enountered
and the message key for the alert is not changed to the desired value.

This change adds a SHA1 key for each change type to the additional info
field, so that Servicenow can add that field to the message key instead
of the full change data. This fixes event grouping.