(MAINT) Fix CI Errors
CI has started to fail because the ruby version in use prior to this
change causes gem dependency conflicts in the Gemfile. Bumping the ruby
version resolves the conflicts.

It is also failing because we have hit the docker cloud rate limit. The
ruby image we use to build the mocking container is been uploaded to
artifactory and we also switch to pulling from artifactory to avoid the
rate limit.