PIE-611 Fix Broken Jenkins Tests
There are some discrepancies between some of the module versions in use
and the reporting_integration module. This current workaround pins the
versions of the provision module and puppet_litmus that work with the
reporting_integration.

Provision module is set to a specific SHA.
Puppet_litmus is set to version 0.26.0 or earlier (lowerbound of 0.18).