(PIE-948) Convert PE setup task to use deploy_pe module
Remove all references to the setup_pe.sh script from the setup_pe task
and replace with a call to the setup_pe method from the deploy_pe
module. Replace all instances of master with server with one exception
for puppet.conf in the test helpers since PE still ships with a master
section in puppet.conf.