PIE-911 Migrate Tests to Cloud CI
We add the CI and Integration Github runners to run unit and acceptance
tests respectively. Removed Jenkinsfile. Removed a test related to the
os.distro fact becasue the new runner uses Centos which does not
natively include that fact. Removed usage of ServiceNow Tasks in the
acceptance tests and replaced with Curl due to networking issues.

PIE-911 Migrate Tests to Cloud CI

We add the CI and Integration Github runners to run unit and acceptance
tests respectively. Removed Jenkinsfile. Removed a test related to the
os.distro fact becasue the new runner uses Centos which does not
natively include that fact. Removed usage of ServiceNow Tasks in the
acceptance tests and replaced with Curl due to networking issues.